requirements.txt
# Flask Framework
flask==3.0.0
flask-cors==4.0.0

# MongoDB
pymongo==4.6.1
dnspython==2.4.2

# Environment Variables
python-dotenv==1.0.0

# Password Hashing
bcrypt==4.1.2

# JWT Authentication
pyjwt==2.8.0

# Date/Time Utilities
python-dateutil==2.8.2
